+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import ButtonBox from './components/buttonbox';
-import Screen from './components/screen';
+import ButtonBox from './components/calculator-components/buttonbox';
+import Screen from './components/calculator-components/screen';
 import Calculator from './components/calculator';
 
 function App() {
